refactor: split largest soft-cap dashboard components (#1036)
Extract sibling modules from the three largest remaining soft-cap
files so each drops under the 600-line watchlist threshold: trove
cell rendering/row-building split out of cdp-trove-table.tsx, the
depositor snapshot table split out of cdp-detail-client.tsx, and the
presentational insight panels split out of v3-flow-insights.tsx. Pure
code moves, no behavior change. Refreshed
docs/notes/file-size-watch.md from scripts/file-size-watchlist.mjs.
